requirements.txt
apache-libcloud==2.8.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
containers==0.0.4
cx-Oracle==7.3.0
dependency-injection==1.2.0
dependency-injector==3.15.6
DynamicMachine==1.0.4
idna==2.9
importlib-metadata==1.6.0
jsonconfigfile==1.0
more-itertools==8.2.0
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pexpect==4.8.0
pluggy==0.13.1
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-repeat==0.8.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
singleton3==1.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
